requirements.txt
# Requirements
gradio==4.44.1
ipywidgets==8.1.3
llama-index-core==0.10.50
llama-index-readers-file==0.1.25
llama-index-llms-nvidia==0.1.3
llama-index-embeddings-nvidia==0.1.4
llama-index-postprocessor-nvidia-rerank==0.1.2
llama-index-embeddings-openai
